# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Quotes API",
    "description": "This is swagger documentation of NodeJS API project of Quotes Application",
    "contact": {
      "email": "pshirsat.cci@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "name": "Authorization",
      "in": "header",
      "description": "Enter the token with the Bearer: prefix, e.g. \"Bearer abcde12345\"",
      "securityScheme": "bearerAuth"
    }
  },
  "paths": {
...
    "/quote/": {
      "get": {
        "tags": [
          "Quotes"
        ],
        "summary": "API to get all the quotes",
        "description": "Find all quotes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Fetched all quotes successfully",
            "schema": {
              "$ref": "#/definitions/QuotesArrayResponse"
            }
          },
          "404": {
            "description": "Quotes not found",
            "schema": {
              "$ref": "#/definitions/NotFoundResponse"
            }
          },
          "500": {
            "description": "Unknown server side error",
            "schema": {
              "$ref": "#/definitions/ServerSideError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Quotes"
        ],
        "summary": "API to create a new quote",
        "description": "Add a new quote",
        "parameters": [
          {
            "name": "Quote",
            "description": "Quote data to create a new quote",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/QuoteRequestData"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Quote created successfully",
            "schema": {
              "$ref": "#/definitions/QuoteResponse"
            }
          },
          "400": {
            "description": "validation error",
            "schema": {
              "$ref": "#/definitions/ValidationResponse"
            }
          },
          "404": {
            "description": "User validation error",
            "schema": {
              "$ref": "#/definitions/ValidationResponse"
            }
          },
          "500": {
            "description": "Unknown server side error",
            "schema": {
              "$ref": "#/definitions/ServerSideError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/author/": {
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Fetch all authors",
        "description": "Fetch all authors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authors data fetched successfully",
            "schema": {
              "$ref": "#/definitions/AuthorResponse"
            }
          },
          "404": {
            "description": "Authors data not found",
            "schema": {
              "$ref": "#/definitions/NotFoundResponse"
            }
          },
          "500": {
            "description": "Unknown server side error",
            "schema": {
              "$ref": "#/definitions/ServerSideError"
            }
          }
        }
      }
    },
    "/user/": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "API to register a new user",
        "description": "Registration of a new user",
        "parameters": [
          {
            "name": "user",
            "description": "user data to store",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created user successfully",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "validation error",
            "schema": {
              "$ref": "#/definitions/ValidationResponse"
            }
          },
          "500": {
            "description": "Unknown server side error",
            "schema": {
              "$ref": "#/definitions/ServerSideError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ValidationResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Not a Valid UUID "
        }
      }
    },
    "QuoteRequestData": {
      "type": "object",
      "properties": {
        "quote": {
          "type": "string",
          "example": "Education is the most powerful weapon that we can use to change the world."
        },
        "author": {
          "type": "string",
          "example": "Nelson Mandela"
        },
        "tags": {
          "type": "string",
          "example": "life; education;"
        }
      }
    },
    "QuotesArrayResponse": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "quote_id": {
            "type": "string",
            "example": "f944af0d-5637-4810-b102-15509cc2c7d8"
          },
          "quote": {
            "type": "string",
            "example": "Technology is just a tool. In terms of getting the kids working together and motivating them, the teacher is most important"
          },
          "author": {
            "type": "string",
            "example": "Bill Gates"
          },
          "likes": {
            "type": "number",
            "example": 10
          },
          "dislikes": {
            "type": "number",
            "example": 3
          },
          "tags": {
            "type": "string",
            "example": "technology;"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-02-08T07:13:44.194Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-02-08T07:13:44.194Z"
          }
        }
      }
    },
    "QuoteResponse": {
      "type": "object",
      "properties": {
        "quote_id": {
          "type": "string",
          "example": "eafaf86e-aa92-4fe7-9d10-86add792a80d"
        },
        "quote": {
          "type": "string",
          "example": "Do not confine your children to your own learning, for they were born in another time."
        },
        "author": {
          "type": "string",
          "example": "Chinese Proverb"
        },
        "likes": {
          "type": "number",
          "example": 0
        },
        "dislikes": {
          "type": "number",
          "example": 0
        },
        "tags": {
          "type": "string",
          "example": "life;"
        },
        "createdAt": {
          "type": "string",
          "example": "2023-02-08T07:13:44.194Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2023-02-08T07:13:44.194Z"
        }
      }
    },
    "AuthorResponse": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "author": {
            "type": "string",
            "example": "Pratap Shirsat"
          }
        }
      }
    },
    "ServerSideError": {
      "type": "object",
      "properties": {
        "Error": {
          "type": "string",
          "example": "Server side error occured"
        }
      }
    },
...
    "NotFoundResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Requested data not found"
        }
      }
    },
    "UserRequest": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string",
              "example": "pratap"
            },
            "password": {
              "type": "string",
              "example": "Cci12345"
            }
          }
        }
      }
    },
    "UserResponse": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "pratap"
        },
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9 <- some auth token"
        }
      }
    },
...
  }
}